Avoid to use writing linked path
Fluentd usually uses predictable paths for dumping object information.

However, if systemd manages fluentd, it will send a SIGCONT to fluentd
when the managed service is stopped or restarted (see 'man systemd.kill').
Thus, if a malicious link(s) are in place in /tmp, the attack will eventually be triggered.

Signed-off-by: Hiroshi Hatake <hatake@calyptia.com>